REFACTOR player movement and camera
- Change everything from float to double

+++ src/app/scenes/Camera.hpp
@@ -13,24 +13,24 @@ class ViewFrustum
 public:
 
 	ViewFrustum(
-		const glm::vec3 & pos, const glm::vec3 & front, const glm::vec3 & up,
-		const float fov, const float ratio, const float nearD, const float farD
+		const glm::dvec3 & pos, const glm::dvec3 & front, const glm::dvec3 & up,
+		const double fov, const double ratio, const double nearD, const double farD
 	);
 
-	bool sphereInFrustum(const glm::vec3 & center, float radius) const;
+	bool sphereInFrustum(const glm::dvec3 & center, double radius) const;
 
-	glm::vec3 nbr, nbl, ntr, ntl, fbr, fbl, ftr, ftl;
+	glm::dvec3 nbr, nbl, ntr, ntl, fbr, fbl, ftr, ftl;
 
 private:
 
-	glm::vec3 m_camera_position;
-	glm::vec3 m_x, m_y, m_z;
-	float m_nearD, m_farD;
-	float m_ratio;
-	float m_fov, m_tang;
-	float m_nearH, m_nearW;
-	float m_farH, m_farW;
-	float m_sphereFactorX, m_sphereFactorY;
+	glm::dvec3 m_camera_position;
+	glm::dvec3 m_x, m_y, m_z;
+	double m_nearD, m_farD;
+	double m_ratio;
+	double m_fov, m_tang;
+	double m_nearH, m_nearW;
+	double m_farH, m_farW;
+	double m_sphereFactorX, m_sphereFactorY;
 };
 
 /**
@@ -45,7 +45,7 @@ public:
 	{
 		glm::mat4 view;
 		glm::mat4 projection;
-		glm::vec3 position;
+		glm::dvec3 position;
 		ViewFrustum view_frustum;
 	};
 
@@ -72,7 +72,7 @@ public:
 	 *
 	 * @param move
 	 */
-	void movePosition(const glm::vec3 & move);
+	void movePosition(const glm::dvec3 & move);
 
 	/**
 	 * @brief Move the camera rotation from the cursor movement.
@@ -80,53 +80,53 @@ public:
 	 * @param x_offset x movement of the cursor.
 	 * @param y_offset y movement of the cursor.
 	 */
-	void moveDirection(float x_offset, float y_offset);
+	void moveDirection(double x_offset, double y_offset);
 
 	/**
 	 * @brief Set the position of the camera.
 	 *
 	 * @param position
 	 */
-	void setPosition(const glm::vec3 & position);
+	void setPosition(const glm::dvec3 & position);
 
 	/**
 	 * @brief Set the pitch and yaw of the camera to look at a target.
 	 *
 	 * @param target
 	 */
-	void lookAt(const glm::vec3 & target);
+	void lookAt(const glm::dvec3 & target);
 
-	glm::vec3 getPosition() const
+	glm::dvec3 getPosition() const
 	{
 		std::lock_guard<std::mutex> lock(m_mutex);
 		return position;
 	}
 
-	float getPitch() const
+	double getPitch() const
 	{
 		std::lock_guard<std::mutex> lock(m_mutex);
 		return pitch;
 	}
 
-	float getYaw() const
+	double getYaw() const
 	{
 		std::lock_guard<std::mutex> lock(m_mutex);
 		return yaw;
 	}
 
-	void setPitch(float pitch)
+	void setPitch(double pitch)
 	{
 		std::lock_guard<std::mutex> lock(m_mutex);
 		this->pitch = pitch;
 	}
 
-	void setYaw(float yaw)
+	void setYaw(double yaw)
 	{
 		std::lock_guard<std::mutex> lock(m_mutex);
 		this->yaw = yaw;
 	}
 
-	RenderInfo getRenderInfo(float aspect_ratio) const
+	RenderInfo getRenderInfo(double aspect_ratio) const
 	{
 		std::lock_guard<std::mutex> lock(m_mutex);
 		return {
@@ -139,23 +139,23 @@ public:
 
 private:
 
-	glm::vec3 position{ 0.0f, 0.0f, 0.0f };
-	float pitch{ 0.0f };
-	float yaw{ 0.0f };
-	glm::vec3 up{ 0.0f, 1.0f, 0.0f };
-	float fov{ 80.0f };
-	float near_plane{ 0.01f };
-	float far_plane{ 1000.0f };
+	glm::dvec3 position{ 0.0f, 0.0f, 0.0f };
+	double pitch{ 0.0f };
+	double yaw{ 0.0f };
+	glm::dvec3 up{ 0.0f, 1.0f, 0.0f };
+	double fov{ 80.0f };
+	double near_plane{ 0.01f };
+	double far_plane{ 1000.0f };
 
 	mutable std::mutex m_mutex;
 
-	glm::vec3 direction() const;
+	glm::dvec3 direction() const;
 
 	glm::mat4 getViewMatrix() const;
 
-	glm::mat4 getProjectionMatrix(float aspect_ratio) const;
+	glm::mat4 getProjectionMatrix(double aspect_ratio) const;
 
-	ViewFrustum getViewFrustum(float aspect_ratio) const;
+	ViewFrustum getViewFrustum(double aspect_ratio) const;
 
 
 };
